Extract helper function from the config pkg
Moving pure helper function to pkg/config/utils allow to use them with comp/core/config comp
