+++ pkg/config/utils/miscellaneous.go
@@ -3,6 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+// Package utils offers a number of high level helpers to work with the configuration
 package utils
 
 import (
